Select trials between first and last spike
Only takes trials into account between the first and the last
spike event. These spike events are adjusted to match the
fixation event and the lever release. Moreover, the timing
of the first spike can be read from the spike times or from
the excel file spike sorting notes. Spike sorting notes are
not up to date for this particular feature yet.

Saving of plotting functions was disabled because Mac at
PNI cannot handle the save function in the script.

v_12 should be adjusted to v_13

 # Please enter the commit message for your changes. Lines starting